Completed Registration Component.
