+++ src/App.css
@@ -1,9 +1,9 @@
-/* Login Input Email and Password Styling */
+/* Sign In Input Username and Password Styling */
 
 input[type="text"],
 input[type="password"] {
   width: 100%;
   margin: 0.2rem;
   background-color: white !important;
-  color: blue !important;
+  color: navy !important;
 }
